fix(node/path): Remove toFileUrl - Remove the toFileUrl function from node/path since it doesn't exist there - Add more assertions against referential invariants of the path export of node - Fixes #18177
