+++ ext/node/polyfills/path/posix.ts
@@ -17,7 +17,6 @@ export const {
   relative,
   resolve,
   sep,
-  toFileUrl,
   toNamespacedPath,
 } = path.posix;
