+++ ext/node/polyfills/path/win32.ts
@@ -17,7 +17,6 @@ export const {
   relative,
   resolve,
   sep,
-  toFileUrl,
   toNamespacedPath,
 } = path.win32;
